pyproject.toml
[project]
name = "cross-dataset-discovery"
version = "0.0.1"
description = "Cross dataset discovery"
authors = [{ name = "Apostolos Koukouvinis", email = "a.koukouvinis@athenarc.gr" }]
readme = "README.md"
keywords = ['python']
requires-python = ">=3.9,<4.0"
classifiers = [
    "Intended Audience :: Developers",
    "Programming Language :: Python",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
    "Topic :: Software Development :: Libraries :: Python Modules",
]
dependencies = [
    "argcomplete",
    "black",
    "brotli",
    "brotlicffi",
    "cairosvg",
    "chardet",
    "chonky",
    "ConfigParser",
    "contextlib2",
    "cryptography",
    "docutils",
    "eval_type_backport",
    "exceptiongroup",
    "faiss_cpu",
    "guidance",
    "h2",
    "HTMLParser",
    "huggingface_hub",
    "jieba",
    "langchain",
    "langchain_community",
    "langchain_core",
    "langchain_ollama",
    "langchain_text_splitters",
    "langdetect",
    "lxml",
    "matplotlib",
    "mock",
    "mxbai_rerank",
    "pandas",
    "pexpect",
    "Pillow",
    "pydantic",
    "pyjnius",
    "pyOpenSSL",
    "pyserini",
    "pytidylib",
    "pytz",
    "regex",
    "ruff",
    "semchunk",
    "sentence_transformers",
    "simplejson",
    "thread",
    "torch",
    "tqdm",
    "transformers",
    "Twisted",
    "typing_extensions",
    "ukkonen",
    #"vllm",
    "zstandard",
    #"ragatouille",
    #"pylate",
    "fastapi",
    "uvicorn[standard]",
    "psycopg2-binary",
    "pgvector",
    "python-dotenv",
    "gunicorn",
]
[project.optional-dependencies]
gpu = [
    "vllm",
    "ragatouille",
    "pylate"
]
dev = [
    "pytest>=7.2.0",
    "pre-commit>=2.20.0",
    "mkdocs-material>=8.5.10",
    "mkdocstrings[python]>=0.26.1"
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.setuptools]
package-dir = { "" = "." }

[tool.setuptools.packages.find]
where   = ["cross_dataset_discovery"]
include = ["*"]

[tool.pytest.ini_options]
testpaths = ["tests"]

[tool.bandit]
skips = ["B301", "B403", "B404"]
